# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "description": "API para el sistema de gestión de Recursos Humanos",
        "contact": {
            "email": "admin@empresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Servidor local alternativo"
        }
    ],
    "paths": {
...
        "/api/tipos-contrato": {
            "get": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Obtener lista de tipos de contrato",
                "description": "Retorna una lista paginada de todos los tipos de contrato",
                "operationId": "105e553e4d21eb59c4c00ab6d284843f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página (máximo 100)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tipos de contrato obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TipoContrato"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Crear un nuevo tipo de contrato",
                "description": "Crea un nuevo tipo de contrato en el sistema",
                "operationId": "ba37d1410db5c1d8db37fb80f768340d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Indefinido"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de contrato creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoContrato"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de contrato creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tipos-contrato/{id}": {
            "get": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Obtener un tipo de contrato específico",
                "description": "Retorna la información de un tipo de contrato por su ID",
                "operationId": "11f17d46bc7c9f5e2ab54ae805386c68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de contrato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoContrato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de contrato no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Actualizar un tipo de contrato",
                "description": "Actualiza un tipo de contrato existente",
                "operationId": "3e0452b9e560344088acdc441ad1a5da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del tipo de contrato",
                                        "type": "string",
                                        "example": "Contrato por Horas"
                                    },
                                    "descripcion": {
                                        "description": "Descripción detallada del tipo de contrato",
                                        "type": "string",
                                        "example": "Contrato para empleados que trabajan por horas"
                                    },
                                    "estado": {
                                        "description": "Estado del tipo de contrato",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de contrato actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de contrato actualizado correctamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoContrato"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de contrato no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo nombre es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Eliminar un tipo de contrato",
                "description": "Elimina un tipo de contrato existente. Nota: Solo se puede eliminar si no tiene empleados asociados.",
                "operationId": "2bbf4616a30e94c5f1e7933075cbef78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de contrato eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de contrato eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de contrato no encontrado"
                    },
                    "409": {
                        "description": "Conflicto - No se puede eliminar porque tiene empleados asociados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se puede eliminar el tipo de contrato porque tiene empleados asociados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TipoContrato": {
                "title": "Tipo Contrato",
                "description": "Modelo de Tipo de Contrato",
                "properties": {
                    "id": {
                        "description": "ID único del tipo de contrato",
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "description": "Nombre del tipo de contrato",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Indefinido"
                    },
                    "estado": {
                        "description": "Estado del tipo de contrato",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}